simplify diff on rawdb.
Extract only celo function to rawdb/celo.go
